refactor: create function checked abs path
